Add files via upload
Using this code, we can generate a sheet consisted of LJ atoms by following an FCC pattern while considering its face (111). Outputs of this code are written in two different file formats, .xyz and .gro, from which we can visualize the sheet and see if PBC is obyed or not.